# src/apis/users/users.controller.ts
import { Body, Controller, Delete, Get, Param, ParseIntPipe, Patch, Post, Session, UseGuards } from '@nestjs/common';
import { CurrentUser } from '../../utils/decorators/current-user.decorator';
import { AuthGuard } from '../../utils/guards/auth.guard';
import { Serialize } from '../../utils/interceptors/serialize.interceptor';
import { AuthService } from './auth.service';
import { CreateUserDto } from './dtos/create-user.dto';
import { GetUsersDto } from './dtos/get-users.dto';
import { SigninDto } from './dtos/signin.dto';
import { UpdateUserDto } from './dtos/update-user.dto';
import { UserDto } from './dtos/user.dto';
import { UsersService } from './users.service';

@Controller('users')
@Serialize(UserDto)
export class UsersController {

    constructor(
        private authService: AuthService,
        private usersService: UsersService
    ) {}

    @Get('whoami')
    @UseGuards(AuthGuard)
    whoami(@CurrentUser() user: string) {
        return user;
    }

    @Post('signup')
    async signup(@Body() body: CreateUserDto, @Session() session: any) {
        const user = await this.authService.signup(body);
        session.userId = user.id;

        return user;
    }

    @Post('signin')
    async signin(@Body() body: SigninDto, @Session() session: any) {
        const user = await this.authService.signin(body);
        session.userId = user.id;

        return user;
    }

    @Post('signout')
    signout(@Session() session: any) {
        session.id = null;
    }

    @Post('create')
    create(@Body() body: CreateUserDto) {
        return this.usersService.create(body);
    }

    @Get('get/:id')
    findOne(@Param('id', ParseIntPipe) id: number) {
        return this.usersService.findOne(id);
    }

    @Get('get')
    findAll(@Body() body: GetUsersDto) {
        return this.usersService.findAll(body);
    }

    @Patch('update/:id')
    update(@Param('id', ParseIntPipe) id: number, @Body() body: UpdateUserDto) {
        return this.usersService.update(id, body);
    }

    @Delete('delete/:id')
    remove(@Param('id', ParseIntPipe) id: number) {
        return this.usersService.remove(id);
    }
}